cotroller连接

+++ questionaire/src/main/java/com/djngroup/questionaire/UserRepository.java
@@ -2,7 +2,9 @@ package com.djngroup.questionaire;
 import org.springframework.data.repository.CrudRepository;
 import com.djngroup.questionaire.entity.User;
 
+import java.util.List;
 
-public interface UserRepository extends CrudRepository<User, Long> {
+
+public interface UserRepository extends CrudRepository<User, Integer> {
 
 }
